[Decomp][ObjectMgr] Cohesion split: extract player create-info + level stats to ObjectMgrPlayerInfo.cpp
