+++ src/game/Object/ObjectMgr.cpp
@@ -2429,870 +2429,11 @@ void ObjectMgr::LoadItemRequiredTarget()
     sLog.outString();
 }
 
-/**
- * @brief Loads pet base stats for each supported level.
- */
-void ObjectMgr::LoadPetLevelInfo()
-{
-    uint32 count = 0;
 
-    // Loading levels data
-    {
-        //                                                 0               1      2   3     4    5    6    7     8    9
-        QueryResult* result  = WorldDatabase.Query("SELECT `creature_entry`, `level`, `hp`, `mana`, `str`, `agi`, `sta`, `inte`, `spi`, `armor` FROM `pet_levelstats`");
 
 
-        if (!result)
-        {
-            BarGoLink bar(1);
-            bar.step();
-            sLog.outString(">> Loaded %u level pet stats definitions", count);
-            sLog.outErrorDb("Error loading `pet_levelstats` table or empty table.");
-            sLog.outString();
-            return;
-        }
 
-        BarGoLink bar(result->GetRowCount());
 
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 creature_id = fields[0].GetUInt32();
-            if (!sCreatureStorage.LookupEntry<CreatureInfo>(creature_id))
-            {
-                sLog.outErrorDb("Wrong creature id %u in `pet_levelstats` table, ignoring.", creature_id);
-                continue;
-            }
-
-            uint32 current_level = fields[1].GetUInt32();
-            if (current_level > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-            {
-                if (current_level > STRONG_MAX_LEVEL)       // hardcoded level maximum
-                {
-                    sLog.outErrorDb("Wrong (> %u) level %u in `pet_levelstats` table, ignoring.", STRONG_MAX_LEVEL, current_level);
-                }
-                else
-                {
-                    DETAIL_FILTER_LOG(LOG_FILTER_DB_STRICTED_CHECK, "Unused (> MaxPlayerLevel in mangosd.conf) level %u in `pet_levelstats` table, ignoring.", current_level);
-                    ++count;                                // make result loading percent "expected" correct in case disabled detail mode for example.
-                }
-                continue;
-            }
-            else if (current_level < 1)
-            {
-                sLog.outErrorDb("Wrong (<1) level %u in `pet_levelstats` table, ignoring.", current_level);
-                continue;
-            }
-
-            PetLevelInfo*& pInfoMapEntry = petInfo[creature_id];
-
-            if (pInfoMapEntry == NULL)
-            {
-                pInfoMapEntry =  new PetLevelInfo[sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL)];
-            }
-
-            // data for level 1 stored in [0] array element, ...
-            PetLevelInfo* pLevelInfo = &pInfoMapEntry[current_level - 1];
-
-            pLevelInfo->health = fields[2].GetUInt16();
-            pLevelInfo->mana   = fields[3].GetUInt16();
-            pLevelInfo->armor  = fields[9].GetUInt16();
-
-            for (int i = 0; i < MAX_STATS; ++i)
-            {
-                pLevelInfo->stats[i] = fields[i + 4].GetUInt16();
-            }
-
-            bar.step();
-            ++count;
-        }
-        while (result->NextRow());
-
-        delete result;
-    }
-
-    // Fill gaps and check integrity
-    for (PetLevelInfoMap::iterator itr = petInfo.begin(); itr != petInfo.end(); ++itr)
-    {
-        PetLevelInfo* pInfo = itr->second;
-
-        // fatal error if no level 1 data
-        if (!pInfo || pInfo[0].health == 0)
-        {
-            sLog.outErrorDb("Creature %u does not have pet stats data for Level 1!", itr->first);
-            Log::WaitBeforeContinueIfNeed();
-            exit(1);
-        }
-
-        // fill level gaps
-        for (uint32 level = 1; level < sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL); ++level)
-        {
-            if (pInfo[level].health == 0)
-            {
-                sLog.outErrorDb("Creature %u has no data for Level %i pet stats data, using data of Level %i.", itr->first, level + 1, level);
-                pInfo[level] = pInfo[level - 1];
-            }
-        }
-    }
-
-    sLog.outString(">> Loaded %u level pet stats definitions", count);
-    sLog.outString();
-}
-
-/**
- * @brief Gets pet level stats for a creature entry and level.
- *
- * @param creature_id The creature or family entry.
- * @param level The requested pet level.
- * @return The matching pet level info, or null if unavailable.
- */
-PetLevelInfo const* ObjectMgr::GetPetLevelInfo(uint32 creature_id, uint32 level) const
-{
-    if (level > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-    {
-        level = sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL);
-    }
-
-    PetLevelInfoMap::const_iterator itr = petInfo.find(creature_id);
-    if (itr == petInfo.end())
-    {
-        return NULL;
-    }
-
-    return &itr->second[level - 1];                         // data for level 1 stored in [0] array element, ...
-}
-
-/**
- * @brief Loads player creation, starting inventory, spells, actions, and XP data.
- */
-void ObjectMgr::LoadPlayerInfo()
-{
-    // Load playercreate
-    {
-        //                                                0     1      2    3     4           5           6
-        QueryResult* result = WorldDatabase.Query("SELECT `race`, `class`, `map`, `zone`, `position_x`, `position_y`, `position_z`, `orientation` FROM `playercreateinfo`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u player create definitions", count);
-            sLog.outErrorDb("Error loading `playercreateinfo` table or empty table.");
-            Log::WaitBeforeContinueIfNeed();
-            exit(1);
-        }
-
-        BarGoLink bar(result->GetRowCount());
-
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 current_race  = fields[0].GetUInt32();
-            uint32 current_class = fields[1].GetUInt32();
-            uint32 mapId         = fields[2].GetUInt32();
-            uint32 areaId        = fields[3].GetUInt32();
-            float  positionX     = fields[4].GetFloat();
-            float  positionY     = fields[5].GetFloat();
-            float  positionZ     = fields[6].GetFloat();
-            float  orientation   = fields[7].GetFloat();
-
-            ChrRacesEntry const* rEntry = sChrRacesStore.LookupEntry(current_race);
-            if (!rEntry || !((1 << (current_race - 1)) & RACEMASK_ALL_PLAYABLE))
-            {
-                sLog.outErrorDb("Wrong race %u in `playercreateinfo` table, ignoring.", current_race);
-                continue;
-            }
-
-            ChrClassesEntry const* cEntry = sChrClassesStore.LookupEntry(current_class);
-            if (!cEntry || !((1 << (current_class - 1)) & CLASSMASK_ALL_PLAYABLE))
-            {
-                sLog.outErrorDb("Wrong class %u in `playercreateinfo` table, ignoring.", current_class);
-                continue;
-            }
-
-            // accept DB data only for valid position (and non instanceable)
-            if (!MapManager::IsValidMapCoord(mapId, positionX, positionY, positionZ, orientation))
-            {
-                sLog.outErrorDb("Wrong home position for class %u race %u pair in `playercreateinfo` table, ignoring.", current_class, current_race);
-                continue;
-            }
-
-            if (sMapStore.LookupEntry(mapId)->Instanceable())
-            {
-                sLog.outErrorDb("Home position in instanceable map for class %u race %u pair in `playercreateinfo` table, ignoring.", current_class, current_race);
-                continue;
-            }
-
-            PlayerInfo* pInfo = &playerInfo[current_race][current_class];
-
-            pInfo->mapId       = mapId;
-            pInfo->areaId      = areaId;
-            pInfo->positionX   = positionX;
-            pInfo->positionY   = positionY;
-            pInfo->positionZ   = positionZ;
-            pInfo->orientation = orientation;
-
-            pInfo->displayId_m = rEntry->model_m;
-            pInfo->displayId_f = rEntry->model_f;
-
-            bar.step();
-            ++count;
-        }
-        while (result->NextRow());
-
-        delete result;
-
-        sLog.outString();
-        sLog.outString(">> Loaded %u player create definitions", count);
-    }
-
-    // Load playercreate items
-    {
-        //                                                0     1      2       3
-        QueryResult* result = WorldDatabase.Query("SELECT `race`, `class`, `itemid`, `amount` FROM `playercreateinfo_item`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            bar.step();
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u custom player create items", count);
-        }
-        else
-        {
-            BarGoLink bar(result->GetRowCount());
-
-            do
-            {
-                Field* fields = result->Fetch();
-
-                uint32 current_race = fields[0].GetUInt32();
-                uint32 current_class = fields[1].GetUInt32();
-
-                ChrRacesEntry const* rEntry = sChrRacesStore.LookupEntry(current_race);
-                if (!rEntry || !((1 << (current_race - 1)) & RACEMASK_ALL_PLAYABLE))
-                {
-                    sLog.outErrorDb("Wrong race %u in `playercreateinfo_item` table, ignoring.", current_race);
-                    continue;
-                }
-
-                ChrClassesEntry const* cEntry = sChrClassesStore.LookupEntry(current_class);
-                if (!cEntry || !((1 << (current_class - 1)) & CLASSMASK_ALL_PLAYABLE))
-                {
-                    sLog.outErrorDb("Wrong class %u in `playercreateinfo_item` table, ignoring.", current_class);
-                    continue;
-                }
-
-                PlayerInfo* pInfo = &playerInfo[current_race][current_class];
-
-                uint32 item_id = fields[2].GetUInt32();
-
-                if (!GetItemPrototype(item_id))
-                {
-                    sLog.outErrorDb("Item id %u (race %u class %u) in `playercreateinfo_item` table but not listed in `item_template`, ignoring.", item_id, current_race, current_class);
-                    continue;
-                }
-
-                uint32 amount  = fields[3].GetUInt32();
-
-                if (!amount)
-                {
-                    sLog.outErrorDb("Item id %u (class %u race %u) have amount==0 in `playercreateinfo_item` table, ignoring.", item_id, current_race, current_class);
-                    continue;
-                }
-
-                pInfo->item.push_back(PlayerCreateInfoItem(item_id, amount));
-
-                bar.step();
-                ++count;
-            }
-            while (result->NextRow());
-
-            delete result;
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u custom player create items", count);
-        }
-    }
-
-    // Load playercreate spells
-    {
-        //                                                0     1      2
-        QueryResult* result = WorldDatabase.Query("SELECT `race`, `class`, `Spell` FROM `playercreateinfo_spell`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u player create spells", count);
-            sLog.outErrorDb("Error loading `playercreateinfo_spell` table or empty table.");
-        }
-        else
-        {
-            BarGoLink bar(result->GetRowCount());
-
-            do
-            {
-                Field* fields = result->Fetch();
-
-                uint32 current_race = fields[0].GetUInt32();
-                uint32 current_class = fields[1].GetUInt32();
-
-                ChrRacesEntry const* rEntry = sChrRacesStore.LookupEntry(current_race);
-                if (!rEntry || !((1 << (current_race - 1)) & RACEMASK_ALL_PLAYABLE))
-                {
-                    sLog.outErrorDb("Wrong race %u in `playercreateinfo_spell` table, ignoring.", current_race);
-                    continue;
-                }
-
-                ChrClassesEntry const* cEntry = sChrClassesStore.LookupEntry(current_class);
-                if (!cEntry || !((1 << (current_class - 1)) & CLASSMASK_ALL_PLAYABLE))
-                {
-                    sLog.outErrorDb("Wrong class %u in `playercreateinfo_spell` table, ignoring.", current_class);
-                    continue;
-                }
-
-                uint32 spell_id = fields[2].GetUInt32();
-                if (!sSpellStore.LookupEntry(spell_id))
-                {
-                    sLog.outErrorDb("Non existing spell %u in `playercreateinfo_spell` table, ignoring.", spell_id);
-                    continue;
-                }
-
-                PlayerInfo* pInfo = &playerInfo[current_race][current_class];
-                pInfo->spell.push_back(spell_id);
-
-                bar.step();
-                ++count;
-            }
-            while (result->NextRow());
-
-            delete result;
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u player create spells", count);
-        }
-    }
-
-    // Load playercreate actions
-    {
-        //                                                0     1      2       3       4
-        QueryResult* result = WorldDatabase.Query("SELECT `race`, `class`, `button`, `action`, `type` FROM `playercreateinfo_action`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u player create actions", count);
-            sLog.outErrorDb("Error loading `playercreateinfo_action` table or empty table.");
-        }
-        else
-        {
-            BarGoLink bar(result->GetRowCount());
-
-            do
-            {
-                Field* fields = result->Fetch();
-
-                uint32 current_race = fields[0].GetUInt32();
-                uint32 current_class = fields[1].GetUInt32();
-
-                ChrRacesEntry const* rEntry = sChrRacesStore.LookupEntry(current_race);
-                if (!rEntry || !((1 << (current_race - 1)) & RACEMASK_ALL_PLAYABLE))
-                {
-                    sLog.outErrorDb("Wrong race %u in `playercreateinfo_action` table, ignoring.", current_race);
-                    continue;
-                }
-
-                ChrClassesEntry const* cEntry = sChrClassesStore.LookupEntry(current_class);
-                if (!cEntry || !((1 << (current_class - 1)) & CLASSMASK_ALL_PLAYABLE))
-                {
-                    sLog.outErrorDb("Wrong class %u in `playercreateinfo_action` table, ignoring.", current_class);
-                    continue;
-                }
-
-                uint8 action_button  = fields[2].GetUInt8();
-                uint32 action = fields[3].GetUInt32();
-                uint8 action_type = fields[4].GetUInt8();
-
-                if (!Player::IsActionButtonDataValid(action_button, action, action_type, NULL))
-                {
-                    continue;
-                }
-
-                PlayerInfo* pInfo = &playerInfo[current_race][current_class];
-                pInfo->action.push_back(PlayerCreateInfoAction(action_button, action, action_type));
-
-                bar.step();
-                ++count;
-            }
-            while (result->NextRow());
-
-            delete result;
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u player create actions", count);
-        }
-    }
-
-    // Loading levels data (class only dependent)
-    {
-        //                                                 0      1      2       3
-        QueryResult* result  = WorldDatabase.Query("SELECT `class`, `level`, `basehp`, `basemana` FROM `player_classlevelstats`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u level health/mana definitions", count);
-            sLog.outErrorDb("Error loading `player_classlevelstats` table or empty table.");
-            Log::WaitBeforeContinueIfNeed();
-            exit(1);
-        }
-
-        BarGoLink bar(result->GetRowCount());
-
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 current_class = fields[0].GetUInt32();
-            if (current_class >= MAX_CLASSES)
-            {
-                sLog.outErrorDb("Wrong class %u in `player_classlevelstats` table, ignoring.", current_class);
-                continue;
-            }
-
-            uint32 current_level = fields[1].GetUInt32();
-            if (current_level == 0)
-            {
-                sLog.outErrorDb("Wrong level %u in `player_classlevelstats` table, ignoring.", current_level);
-                continue;
-            }
-            else if (current_level > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-            {
-                if (current_level > STRONG_MAX_LEVEL)       // hardcoded level maximum
-                {
-                    sLog.outErrorDb("Wrong (> %u) level %u in `player_classlevelstats` table, ignoring.", STRONG_MAX_LEVEL, current_level);
-                }
-                else
-                {
-                    DETAIL_FILTER_LOG(LOG_FILTER_DB_STRICTED_CHECK, "Unused (> MaxPlayerLevel in mangosd.conf) level %u in `player_classlevelstats` table, ignoring.", current_level);
-                    ++count;                                // make result loading percent "expected" correct in case disabled detail mode for example.
-                }
-                continue;
-            }
-
-            PlayerClassInfo* pClassInfo = &playerClassInfo[current_class];
-
-            if (!pClassInfo->levelInfo)
-            {
-                pClassInfo->levelInfo = new PlayerClassLevelInfo[sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL)];
-            }
-
-            PlayerClassLevelInfo* pClassLevelInfo = &pClassInfo->levelInfo[current_level - 1];
-
-            pClassLevelInfo->basehealth = fields[2].GetUInt16();
-            pClassLevelInfo->basemana   = fields[3].GetUInt16();
-
-            bar.step();
-            ++count;
-        }
-        while (result->NextRow());
-
-        delete result;
-
-        sLog.outString();
-        sLog.outString(">> Loaded %u level health/mana definitions", count);
-    }
-
-    // Fill gaps and check integrity
-    for (int class_ = 0; class_ < MAX_CLASSES; ++class_)
-    {
-        // skip nonexistent classes
-        if (!sChrClassesStore.LookupEntry(class_))
-        {
-            continue;
-        }
-
-        PlayerClassInfo* pClassInfo = &playerClassInfo[class_];
-
-        // fatal error if no level 1 data
-        if (!pClassInfo->levelInfo || pClassInfo->levelInfo[0].basehealth == 0)
-        {
-            sLog.outErrorDb("Class %i Level 1 does not have health/mana data!", class_);
-            Log::WaitBeforeContinueIfNeed();
-            exit(1);
-        }
-
-        // fill level gaps
-        for (uint32 level = 1; level < sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL); ++level)
-        {
-            if (pClassInfo->levelInfo[level].basehealth == 0)
-            {
-                sLog.outErrorDb("Class %i Level %i does not have health/mana data. Using stats data of level %i.", class_, level + 1, level);
-                pClassInfo->levelInfo[level] = pClassInfo->levelInfo[level - 1];
-            }
-        }
-    }
-
-    // Loading levels data (class/race dependent)
-    {
-        //                                                 0     1      2      3    4    5    6    7
-        QueryResult* result  = WorldDatabase.Query("SELECT `race`, `class`, `level`, `str`, `agi`, `sta`, `inte`, `spi` FROM `player_levelstats`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u level stats definitions", count);
-            sLog.outErrorDb("Error loading `player_levelstats` table or empty table.");
-            Log::WaitBeforeContinueIfNeed();
-            exit(1);
-        }
-
-        BarGoLink bar(result->GetRowCount());
-
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 current_race = fields[0].GetUInt32();
-            uint32 current_class = fields[1].GetUInt32();
-
-            ChrRacesEntry const* rEntry = sChrRacesStore.LookupEntry(current_race);
-            if (!rEntry || !((1 << (current_race - 1)) & RACEMASK_ALL_PLAYABLE))
-            {
-                sLog.outErrorDb("Wrong race %u in `player_levelstats` table, ignoring.", current_race);
-                continue;
-            }
-
-            ChrClassesEntry const* cEntry = sChrClassesStore.LookupEntry(current_class);
-            if (!cEntry || !((1 << (current_class - 1)) & CLASSMASK_ALL_PLAYABLE))
-            {
-                sLog.outErrorDb("Wrong class %u in `player_levelstats` table, ignoring.", current_class);
-                continue;
-            }
-
-            uint32 current_level = fields[2].GetUInt32();
-            if (current_level > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-            {
-                if (current_level > STRONG_MAX_LEVEL)       // hardcoded level maximum
-                {
-                    sLog.outErrorDb("Wrong (> %u) level %u in `player_levelstats` table, ignoring.", STRONG_MAX_LEVEL, current_level);
-                }
-                else
-                {
-                    DETAIL_FILTER_LOG(LOG_FILTER_DB_STRICTED_CHECK, "Unused (> MaxPlayerLevel in mangosd.conf) level %u in `player_levelstats` table, ignoring.", current_level);
-                    ++count;                                // make result loading percent "expected" correct in case disabled detail mode for example.
-                }
-                continue;
-            }
-
-            PlayerInfo* pInfo = &playerInfo[current_race][current_class];
-
-            if (!pInfo->levelInfo)
-            {
-                pInfo->levelInfo = new PlayerLevelInfo[sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL)];
-            }
-
-            PlayerLevelInfo* pLevelInfo = &pInfo->levelInfo[current_level - 1];
-
-            for (int i = 0; i < MAX_STATS; ++i)
-            {
-                pLevelInfo->stats[i] = fields[i + 3].GetUInt8();
-            }
-
-            bar.step();
-            ++count;
-        }
-        while (result->NextRow());
-
-        delete result;
-
-        sLog.outString();
-        sLog.outString(">> Loaded %u level stats definitions", count);
-    }
-
-    // Fill gaps and check integrity
-    for (int race = 0; race < MAX_RACES; ++race)
-    {
-        // skip nonexistent races
-        if (!((1 << (race - 1)) & RACEMASK_ALL_PLAYABLE) || !sChrRacesStore.LookupEntry(race))
-        {
-            continue;
-        }
-
-        for (int class_ = 0; class_ < MAX_CLASSES; ++class_)
-        {
-            // skip nonexistent classes
-            if (!((1 << (class_ - 1)) & CLASSMASK_ALL_PLAYABLE) || !sChrClassesStore.LookupEntry(class_))
-            {
-                continue;
-            }
-
-            PlayerInfo* pInfo = &playerInfo[race][class_];
-
-            // skip non loaded combinations
-            if (!pInfo->displayId_m || !pInfo->displayId_f)
-            {
-                continue;
-            }
-
-            // skip expansion races if not playing with expansion
-            if (sWorld.getConfig(CONFIG_UINT32_EXPANSION) < EXPANSION_TBC && (race == RACE_BLOODELF || race == RACE_DRAENEI))
-            {
-                continue;
-            }
-
-            // fatal error if no level 1 data
-            if (!pInfo->levelInfo || pInfo->levelInfo[0].stats[0] == 0)
-            {
-                sLog.outErrorDb("Race %i Class %i Level 1 does not have stats data!", race, class_);
-                Log::WaitBeforeContinueIfNeed();
-                exit(1);
-            }
-
-            // fill level gaps
-            for (uint32 level = 1; level < sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL); ++level)
-            {
-                if (pInfo->levelInfo[level].stats[0] == 0)
-                {
-                    sLog.outErrorDb("Race %i Class %i Level %i does not have stats data. Using stats data of level %i.", race, class_, level + 1, level);
-                    pInfo->levelInfo[level] = pInfo->levelInfo[level - 1];
-                }
-            }
-        }
-    }
-
-    // Loading xp per level data
-    {
-        mPlayerXPperLevel.resize(sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL));
-        for (uint32 level = 0; level < sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL); ++level)
-        {
-            mPlayerXPperLevel[level] = 0;
-        }
-
-        //                                                 0    1
-        QueryResult* result  = WorldDatabase.Query("SELECT `lvl`, `xp_for_next_level` FROM `player_xp_for_level`");
-
-        uint32 count = 0;
-
-        if (!result)
-        {
-            BarGoLink bar(1);
-
-            sLog.outString();
-            sLog.outString(">> Loaded %u xp for level definitions", count);
-            sLog.outErrorDb("Error loading `player_xp_for_level` table or empty table.");
-            Log::WaitBeforeContinueIfNeed();
-            exit(1);
-        }
-
-        BarGoLink bar(result->GetRowCount());
-
-        do
-        {
-            Field* fields = result->Fetch();
-
-            uint32 current_level = fields[0].GetUInt32();
-            uint32 current_xp    = fields[1].GetUInt32();
-
-            if (current_level >= sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-            {
-                if (current_level > STRONG_MAX_LEVEL)       // hardcoded level maximum
-                {
-                    sLog.outErrorDb("Wrong (> %u) level %u in `player_xp_for_level` table, ignoring.", STRONG_MAX_LEVEL, current_level);
-                }
-                else
-                {
-                    DETAIL_FILTER_LOG(LOG_FILTER_DB_STRICTED_CHECK, "Unused (> MaxPlayerLevel in mangosd.conf) level %u in `player_xp_for_levels` table, ignoring.", current_level);
-                    ++count;                                // make result loading percent "expected" correct in case disabled detail mode for example.
-                }
-                continue;
-            }
-            // PlayerXPperLevel
-            mPlayerXPperLevel[current_level] = current_xp;
-            bar.step();
-            ++count;
-        }
-        while (result->NextRow());
-
-        delete result;
-
-        sLog.outString();
-        sLog.outString(">> Loaded %u xp for level definitions", count);
-    }
-
-    // fill level gaps
-    for (uint32 level = 1; level < sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL); ++level)
-    {
-        if (mPlayerXPperLevel[level] == 0)
-        {
-            sLog.outErrorDb("Level %i does not have XP for level data. Using data of level [%i] + 100.", level + 1, level);
-            mPlayerXPperLevel[level] = mPlayerXPperLevel[level - 1] + 100;
-        }
-    }
-}
-
-/**
- * @brief Gets class-based level info for a player class and level.
- *
- * @param class_ The player class id.
- * @param level The player level.
- * @param info Receives the class level info.
- */
-void ObjectMgr::GetPlayerClassLevelInfo(uint32 class_, uint32 level, PlayerClassLevelInfo* info) const
-{
-    if (level < 1 || class_ >= MAX_CLASSES)
-    {
-        return;
-    }
-
-    PlayerClassInfo const* pInfo = &playerClassInfo[class_];
-
-    if (level > sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-    {
-        level = sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL);
-    }
-
-    *info = pInfo->levelInfo[level - 1];
-}
-
-/**
- * @brief Gets full player level info for a race, class, and level.
- *
- * @param race The player race id.
- * @param class_ The player class id.
- * @param level The requested player level.
- * @param info Receives the computed level info.
- */
-void ObjectMgr::GetPlayerLevelInfo(uint32 race, uint32 class_, uint32 level, PlayerLevelInfo* info) const
-{
-    if (level < 1 || race   >= MAX_RACES || class_ >= MAX_CLASSES)
-    {
-        return;
-    }
-
-    PlayerInfo const* pInfo = &playerInfo[race][class_];
-    if (pInfo->displayId_m == 0 || pInfo->displayId_f == 0)
-    {
-        return;
-    }
-
-    if (level <= sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL))
-    {
-        *info = pInfo->levelInfo[level - 1];
-    }
-    else
-    {
-        BuildPlayerLevelInfo(race, class_, level, info);
-    }
-}
-
-/**
- * @brief Builds extrapolated player level stats beyond the stored base tables.
- *
- * @param race The player race id.
- * @param _class The player class id.
- * @param level The target player level.
- * @param info Receives the computed level info.
- */
-void ObjectMgr::BuildPlayerLevelInfo(uint8 race, uint8 _class, uint8 level, PlayerLevelInfo* info) const
-{
-    // base data (last known level)
-    *info = playerInfo[race][_class].levelInfo[sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL) - 1];
-
-    for (int lvl = sWorld.getConfig(CONFIG_UINT32_MAX_PLAYER_LEVEL) - 1; lvl < level; ++lvl)
-    {
-        switch (_class)
-        {
-            case CLASS_WARRIOR:
-                info->stats[STAT_STRENGTH]  += (lvl > 23 ? 2 : (lvl > 1  ? 1 : 0));
-                info->stats[STAT_STAMINA]   += (lvl > 23 ? 2 : (lvl > 1  ? 1 : 0));
-                info->stats[STAT_AGILITY]   += (lvl > 36 ? 1 : (lvl > 6 && (lvl % 2) ? 1 : 0));
-                info->stats[STAT_INTELLECT] += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_SPIRIT]    += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                break;
-            case CLASS_PALADIN:
-                info->stats[STAT_STRENGTH]  += (lvl > 3  ? 1 : 0);
-                info->stats[STAT_STAMINA]   += (lvl > 33 ? 2 : (lvl > 1 ? 1 : 0));
-                info->stats[STAT_AGILITY]   += (lvl > 38 ? 1 : (lvl > 7 && !(lvl % 2) ? 1 : 0));
-                info->stats[STAT_INTELLECT] += (lvl > 6 && (lvl % 2) ? 1 : 0);
-                info->stats[STAT_SPIRIT]    += (lvl > 7 ? 1 : 0);
-                break;
-            case CLASS_HUNTER:
-                info->stats[STAT_STRENGTH]  += (lvl > 4  ? 1 : 0);
-                info->stats[STAT_STAMINA]   += (lvl > 4  ? 1 : 0);
-                info->stats[STAT_AGILITY]   += (lvl > 33 ? 2 : (lvl > 1 ? 1 : 0));
-                info->stats[STAT_INTELLECT] += (lvl > 8 && (lvl % 2) ? 1 : 0);
-                info->stats[STAT_SPIRIT]    += (lvl > 38 ? 1 : (lvl > 9 && !(lvl % 2) ? 1 : 0));
-                break;
-            case CLASS_ROGUE:
-                info->stats[STAT_STRENGTH]  += (lvl > 5  ? 1 : 0);
-                info->stats[STAT_STAMINA]   += (lvl > 4  ? 1 : 0);
-                info->stats[STAT_AGILITY]   += (lvl > 16 ? 2 : (lvl > 1 ? 1 : 0));
-                info->stats[STAT_INTELLECT] += (lvl > 8 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_SPIRIT]    += (lvl > 38 ? 1 : (lvl > 9 && !(lvl % 2) ? 1 : 0));
-                break;
-            case CLASS_PRIEST:
-                info->stats[STAT_STRENGTH]  += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_STAMINA]   += (lvl > 5  ? 1 : 0);
-                info->stats[STAT_AGILITY]   += (lvl > 38 ? 1 : (lvl > 8 && (lvl % 2) ? 1 : 0));
-                info->stats[STAT_INTELLECT] += (lvl > 22 ? 2 : (lvl > 1 ? 1 : 0));
-                info->stats[STAT_SPIRIT]    += (lvl > 3  ? 1 : 0);
-                break;
-            case CLASS_SHAMAN:
-                info->stats[STAT_STRENGTH]  += (lvl > 34 ? 1 : (lvl > 6 && (lvl % 2) ? 1 : 0));
-                info->stats[STAT_STAMINA]   += (lvl > 4 ? 1 : 0);
-                info->stats[STAT_AGILITY]   += (lvl > 7 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_INTELLECT] += (lvl > 5 ? 1 : 0);
-                info->stats[STAT_SPIRIT]    += (lvl > 4 ? 1 : 0);
-                break;
-            case CLASS_MAGE:
-                info->stats[STAT_STRENGTH]  += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_STAMINA]   += (lvl > 5  ? 1 : 0);
-                info->stats[STAT_AGILITY]   += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_INTELLECT] += (lvl > 24 ? 2 : (lvl > 1 ? 1 : 0));
-                info->stats[STAT_SPIRIT]    += (lvl > 33 ? 2 : (lvl > 2 ? 1 : 0));
-                break;
-            case CLASS_WARLOCK:
-                info->stats[STAT_STRENGTH]  += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_STAMINA]   += (lvl > 38 ? 2 : (lvl > 3 ? 1 : 0));
-                info->stats[STAT_AGILITY]   += (lvl > 9 && !(lvl % 2) ? 1 : 0);
-                info->stats[STAT_INTELLECT] += (lvl > 33 ? 2 : (lvl > 2 ? 1 : 0));
-                info->stats[STAT_SPIRIT]    += (lvl > 38 ? 2 : (lvl > 3 ? 1 : 0));
-                break;
-            case CLASS_DRUID:
-                info->stats[STAT_STRENGTH]  += (lvl > 38 ? 2 : (lvl > 6 && (lvl % 2) ? 1 : 0));
-                info->stats[STAT_STAMINA]   += (lvl > 32 ? 2 : (lvl > 4 ? 1 : 0));
-                info->stats[STAT_AGILITY]   += (lvl > 38 ? 2 : (lvl > 8 && (lvl % 2) ? 1 : 0));
-                info->stats[STAT_INTELLECT] += (lvl > 38 ? 3 : (lvl > 4 ? 1 : 0));
-                info->stats[STAT_SPIRIT]    += (lvl > 38 ? 3 : (lvl > 5 ? 1 : 0));
-        }
-    }
-}
 
 /* ********************************************************************************************* */
 /* *                                Static Wrappers                                              */
